+++ CHANGELOG.md
@@ -98,6 +98,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Arrange proper failover triggering: don't miss events, don't trigger
   if nothing changed.
 
+- Box errors handling in remote-control.
+
 ## [1.2.0] - 2019-10-21
 
 ### Added